Refactor header files so project makes properly
