+++ update.h
@@ -11,8 +11,3 @@ struct Update{
     long delta;
     friend std::ostream& operator<< (std::ostream &out, const Update &update);
 };
-
-std::ostream& operator<< (std::ostream &out, const Update &update){
-    out << "Index: " << update.index << " Value: " << update.delta;
-    return out;
-}
